+++ CHANGELOG.md
@@ -29,7 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -
 
 
-## [1.0.1] - 2023-07-25
+## [1.0.1] - 2023-07-26
 
 
 ### Added